Separate class for the utils for parser
