+++ src/main/scala/com/databricks/spark/xml/parsers/StaxXmlParser.scala
@@ -18,7 +18,7 @@ package com.databricks.spark.xml.parsers
 import java.io.ByteArrayInputStream
 import javax.xml.stream.events.{Attribute, XMLEvent}
 import javax.xml.stream.events._
-import javax.xml.stream.{XMLStreamException, XMLEventReader, XMLInputFactory}
+import javax.xml.stream.{XMLStreamException, XMLStreamConstants, XMLEventReader, XMLInputFactory}
 
 import scala.collection.immutable.TreeMap
 import scala.collection.mutable.ArrayBuffer
@@ -51,7 +51,7 @@ private[xml] object StaxXmlParser {
         val parser = factory.createXMLEventReader(reader)
         try {
           val rootAttributes = {
-            val rootEvent = skipUntil(parser, classOf[StartElement])
+            val rootEvent = StaxXmlParserUtils.skipUntil(parser, XMLStreamConstants.START_ELEMENT)
             rootEvent.asStartElement.getAttributes
               .map(_.asInstanceOf[Attribute]).toArray
           }
@@ -73,31 +73,6 @@ private[xml] object StaxXmlParser {
         }
       }
     }
-  }
-
-  /**
-   * Skips elements until this meets the given type of a element
-   */
-  def skipUntil(parser: XMLEventReader, eventType: Class[_ <: XMLEvent]): XMLEvent = {
-    var event = parser.nextEvent
-    while(parser.hasNext && event.getClass != eventType) {
-      event = parser.nextEvent
-    }
-    event
-  }
-
-  /**
-   * Read data for all continuous character events.
-   */
-  def readDataFully(parser: XMLEventReader): String = {
-    var event = parser.peek
-    var data: String = if (event.isCharacters) "" else null
-    while(event.isCharacters) {
-      data += event.asCharacters.getData
-      parser.nextEvent
-      event = parser.peek
-    }
-    data
   }
 
   /**
@@ -154,12 +129,13 @@ private[xml] object StaxXmlParser {
         (next, dataType) match {
           case (_: EndElement, _) => if (options.treatEmptyValuesAsNulls) null else data
           case (_: StartElement, dt: DataType) => convertComplicatedType(dt)
-          case (_: Characters, dt: DataType) => convertStringTo(readDataFully(parser), dt)
+          case (_: Characters, dt: DataType) =>
+            convertStringTo(StaxXmlParserUtils.readDataFully(parser), dt)
         }
       case (c: Characters, ArrayType(st, _)) if !c.isIgnorableWhiteSpace && !c.isWhiteSpace =>
-        convertStringTo(readDataFully(parser), st)
+        convertStringTo(StaxXmlParserUtils.readDataFully(parser), st)
       case (c: Characters, dt: DataType) if !c.isIgnorableWhiteSpace && !c.isWhiteSpace =>
-        convertStringTo(readDataFully(parser), dt)
+        convertStringTo(StaxXmlParserUtils.readDataFully(parser), dt)
       case (e: XMLEvent, dt: DataType) =>
         sys.error(s"Failed to parse a value for data type $dt with event ${e.toString}")
     }
@@ -317,3 +293,4 @@ private[xml] object StaxXmlParser {
     Row.fromSeq(row)
   }
 }
+
