+++ src/main/scala/com/databricks/spark/xml/util/InferSchema.scala
@@ -17,8 +17,9 @@ package com.databricks.spark.xml.util
 
 import java.io.ByteArrayInputStream
 import javax.xml.stream.events._
-import javax.xml.stream.{XMLStreamException, XMLEventReader, XMLInputFactory}
+import javax.xml.stream.{XMLStreamConstants, XMLStreamException, XMLEventReader, XMLInputFactory}
 
+import com.databricks.spark.xml.parsers.StaxXmlParserUtils
 import org.slf4j.LoggerFactory
 
 import scala.collection.Seq
@@ -88,7 +89,7 @@ private[xml] object InferSchema {
         val parser = factory.createXMLEventReader(reader)
         try {
           val rootAttributes = {
-            val rootEvent = skipUntil(parser, classOf[StartElement])
+            val rootEvent = StaxXmlParserUtils.skipUntil(parser, XMLStreamConstants.START_ELEMENT)
             rootEvent.asStartElement.getAttributes
               .map(_.asInstanceOf[Attribute]).toArray
           }
@@ -140,11 +141,11 @@ private[xml] object InferSchema {
           case _: EndElement if options.treatEmptyValuesAsNulls => NullType
           case _: EndElement => StringType
           case _: StartElement => inferObject(parser, options)
-          case _: Characters => inferTypeFromString(readDataFully(parser))
+          case _: Characters => inferTypeFromString(StaxXmlParserUtils.readDataFully(parser))
         }
       case c: Characters if !c.isIgnorableWhiteSpace && !c.isWhiteSpace =>
         // This means data exists
-        inferTypeFromString(readDataFully(parser))
+        inferTypeFromString(StaxXmlParserUtils.readDataFully(parser))
 
       case e: XMLEvent =>
         sys.error(s"Failed to parse data with unexpected event ${e.toString}")
